Extract types and helpers into domain module
